refactor: externalize dashboard types to resolve HMR issues and clean up provider
