+++ frontend/src/components/DashboardProvider.tsx
@@ -7,29 +7,7 @@ type DashboardProviderTypes = {
     children: React.ReactNode;
 }
 
-export interface SavedLink {
-  id: number;
-  label: string;
-  url: string;
-}
-
-type DashboardContextType = {
-  applications: Application[];
-  setApplications: React.Dispatch<React.SetStateAction<Application[]>>;
-  analytics: AnalyticsResponse | null; 
-  insights: InsightsResponse | null;
-  isLoading: boolean;
-  savedLinks: SavedLink[];
-  setSavedLinks: React.Dispatch<React.SetStateAction<SavedLink[]>>;
-  fetchData: (showLoading?: boolean) => Promise<void>;
-  saveApplication: (data: Partial<Application>, id?: number) => Promise<void>;
-  deleteApplication: (id: number) => Promise<void>;
-  changeApplicationStatus: (id: number, status: ApplicationStatus) => Promise<void>;
-  page: number;
-  hasMore: boolean;
-  isLoadingMore: boolean;
-  loadMoreApplications: () => Promise<void>;
-};
+import type { DashboardContextType, SavedLink } from '../types/dashboard';
 
 const DashboardContext = createContext<DashboardContextType | null>(null);
